refactor: convert Grade to enum
