+++ src/api/index.ts
@@ -1,3 +1,2 @@
-export type { Submission } from "./services/common/types";
+export type { Grade, Submission } from "./services/common/types";
 export { getMembers } from "./services/store/storeService";
-export type { Grade } from "./types";
