--- src/api/types.ts
@@ -1,8 +1,0 @@
-export type Grade = "SEED" | "SPROUT" | "SMALL_TREE" | "BIG_TREE";
-
-export const Grades = {
-  SEED: "SEED",
-  SPROUT: "SPROUT",
-  SMALL_TREE: "SMALL_TREE",
-  BIG_TREE: "BIG_TREE",
-} as const;
